remove ShortCircuitAnalysisService
