+++ src/main/java/org/gridsuite/study/server/service/NotificationService.java
@@ -42,14 +42,15 @@ public class NotificationService {
     public static final String UPDATE_TYPE_LINE = "line";
     public static final String UPDATE_TYPE_LOADFLOW = "loadflow";
     public static final String UPDATE_TYPE_LOADFLOW_STATUS = "loadflow_status";
-    public static final String UPDATE_TYPE_SHORT_CIRCUIT = "shortcircuit";
-    public static final String UPDATE_TYPE_SHORT_CIRCUIT_STATUS = "shortcircuit_status";
     public static final String UPDATE_TYPE_SECURITY_ANALYSIS_FAILED = "securityAnalysis_failed";
     public static final String UPDATE_TYPE_SECURITY_ANALYSIS_RESULT = "securityAnalysisResult";
     public static final String UPDATE_TYPE_SECURITY_ANALYSIS_STATUS = "securityAnalysis_status";
     public static final String UPDATE_TYPE_SENSITIVITY_ANALYSIS_RESULT = "sensitivityAnalysisResult";
     public static final String UPDATE_TYPE_SENSITIVITY_ANALYSIS_STATUS = "sensitivityAnalysis_status";
     public static final String UPDATE_TYPE_SENSITIVITY_ANALYSIS_FAILED = "sensitivityAnalysis_failed";
+    public static final String UPDATE_TYPE_SHORT_CIRCUIT_RESULT = "shortCircuitAnalysisResult";
+    public static final String UPDATE_TYPE_SHORT_CIRCUIT_STATUS = "shortCircuitAnalysis_status";
+    public static final String UPDATE_TYPE_SHORT_CIRCUIT_FAILED = "shortCircuitAnalysis_failed";
     public static final String UPDATE_TYPE_STUDIES = "studies";
     public static final String UPDATE_TYPE_STUDY = "study";
     public static final String UPDATE_TYPE_STUDY_DELETE = "deleteStudy";
